feat(commonMain/recipes): implemented fractional servings (resolved #143)

+++ composeApp/src/commonMain/kotlin/de/kitshn/api/tandoor/model/recipe/TandoorRecipe.kt
@@ -151,7 +151,7 @@ class TandoorRecipe(
     suspend fun shopping(
         ingredients: List<Int>? = null,
         mealplan: TandoorMealPlan? = null,
-        servings: Int? = null
+        servings: Double? = null
     ) {
         if(client == null) return
